pyproject.toml
[project]
name = "update-resume"
version = "0.1.0"
description = ""
authors = [
    {name = "error-harry",email = "nargideharsh@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "playwright (>=1.58.0,<2.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
